Healthcare-System-master/healthcare-backend/requirements.txt
# ============================================
# Healthcare Backend - Python Dependencies
# ============================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# MongoDB
motor==3.7.0
pymongo==4.10.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# CORS & Security
python-multipart==0.0.18

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# WebSocket Support (included in FastAPI)
websockets==14.2

# Date/Time utilities
python-dateutil==2.9.0

# ML & Embeddings
sentence-transformers==3.3.1
torch==2.4.1
